rename: browserUtils -> webUtils
