+++ packages/toolkit/src/extensionShared.ts
@@ -33,7 +33,7 @@ import { CredentialsStore } from './auth/credentials/store'
 import { initializeAwsCredentialsStatusBarItem } from './auth/ui/statusBarItem'
 import { RegionProvider } from './shared/regions/regionProvider'
 import { ChildProcess } from './shared/utilities/childProcess'
-import { isWeb } from './common/browserUtils'
+import { isWeb } from './common/webUtils'
 import { registerErrorHandler as registerCommandErrorHandler } from './shared/vscode/commands2'
 import { ToolkitError, isUserCancelledError, resolveErrorMessageToDisplay } from './shared/errors'
 import { getLogger } from './shared/logger'
